refactor: move helper classes to separate files (#46)
