+++ src/main/java/org/apache/tinkerpop/gremlin/tinkergraph/structure/OverflowDbNode.java
@@ -430,55 +430,6 @@ public abstract class OverflowDbNode implements Vertex {
     }
   }
 
-  private static class DummyEdgeIterator implements Iterator<Edge> {
-    private final Object[] array;
-    private int current;
-    private final int begin;
-    private final int exclusiveEnd;
-    private final int strideSize;
-    private final Direction direction;
-    private final String label;
-    private final VertexRef<OverflowDbNode> thisRef;
-
-    DummyEdgeIterator(Object[] array, int begin, int exclusiveEnd, int strideSize,
-                      Direction direction, String label, VertexRef<OverflowDbNode> thisRef) {
-      this.array = array;
-      this.begin = begin;
-      this.current = begin;
-      this.exclusiveEnd = exclusiveEnd;
-      this.strideSize = strideSize;
-      this.direction = direction;
-      this.label = label;
-      this.thisRef = thisRef;
-    }
-
-    @Override
-    public boolean hasNext() {
-      /* there may be holes, e.g. if an edge was removed */
-      while (current < exclusiveEnd && array[current] == null) {
-        current += strideSize;
-      }
-      return current < exclusiveEnd;
-    }
-
-    @Override
-    public Edge next() {
-      if (!hasNext()) throw new NoSuchElementException();
-
-      VertexRef<OverflowDbNode> otherRef = (VertexRef<OverflowDbNode>) array[current];
-      OverflowDbEdge dummyEdge;
-      if (direction == Direction.OUT) {
-        dummyEdge = thisRef.element.instantiateDummyEdge(label, thisRef, otherRef);
-        dummyEdge.setOutBlockOffset(current - begin);
-      } else {
-        dummyEdge = thisRef.element.instantiateDummyEdge(label, otherRef, thisRef);
-        dummyEdge.setInBlockOffset(current - begin);
-      }
-      current += strideSize;
-      return dummyEdge;
-    }
-  }
-
   private Iterator<VertexRef> createAdjacentVertexIterator(Direction direction, String label) {
     int offsetPos = getPositionInEdgeOffsets(direction, label);
     if (offsetPos != -1) {
@@ -489,32 +440,6 @@ public abstract class OverflowDbNode implements Vertex {
       return new ArrayOffsetIterator<>(adjacentVerticesWithProperties, start, start + length, strideSize);
     } else {
       return Collections.emptyIterator();
-    }
-  }
-
-  private static class ArrayOffsetIterator<T> implements Iterator<T>  {
-    private final Object[] array;
-    private int current;
-    private final int exclusiveEnd;
-    private final int strideSize;
-
-    ArrayOffsetIterator(Object[] array, int begin, int exclusiveEnd, int strideSize) {
-      this.array = array;
-      this.current = begin;
-      this.exclusiveEnd = exclusiveEnd;
-      this.strideSize = strideSize;
-    }
-
-    @Override
-    public boolean hasNext() {
-      return current < exclusiveEnd;
-    }
-
-    @Override
-    public T next() {
-      T element = (T) array[current];
-      current += strideSize;
-      return element;
     }
   }
